Fix a few compiler warnings

+++ Kerberos/src/Kerberos/Debug/Instrumentor.h
@@ -110,7 +110,7 @@ namespace Kerberos
 			const long long start = std::chrono::time_point_cast<std::chrono::microseconds>(m_StartTimepoint).time_since_epoch().count();
 			const long long end = std::chrono::time_point_cast<std::chrono::microseconds>(endTimepoint).time_since_epoch().count();
 
-			const uint32_t threadID = std::hash<std::thread::id>{}(std::this_thread::get_id());
+			const uint32_t threadID = static_cast<uint32_t>(std::hash<std::thread::id>{}(std::this_thread::get_id()));
 			Instrumentor::Get().WriteProfile({ .Name= m_Name, .Start= start, .End= end, .ThreadID= threadID});
 
 			m_Stopped = true;
